cuter nav for bridge between deposit and withdraw
