+++ src/app/old-layout-client.tsx
@@ -5,6 +5,18 @@ import { BridgeConfig } from "@/util/atoms";
 import Header from "@/comps/Header";
 
 import LayoutClient from "./layout-client";
+import { NavTabs } from "./reskin/components/tabs/nav-tabs";
+
+const tabs = [
+  {
+    label: "Deposit",
+    link: "/",
+  },
+  {
+    label: "Withdraw",
+    link: "/withdraw",
+  },
+];
 
 export default function OldLayoutClient({
   children,
@@ -17,8 +29,10 @@ export default function OldLayoutClient({
 }>) {
   return (
     <LayoutClient config={config}>
-      <main className="min-w-screen bg-white  flex items-center flex-col min-h-screen ">
+      <main className="min-w-screen bg-white  flex items-center flex-col min-h-screen w-full">
         <Header config={config} />
+        <hr className="bg-black hidden w-full sm:block" />
+        <NavTabs tabs={tabs} />
         {children}
         <a
           key="faqs"
@@ -30,6 +44,8 @@ export default function OldLayoutClient({
           FAQs
         </a>
         {footer}
+        {/* to fill the place of the nav tabs in mobile */}
+        <hr className="block w-full sm:hidden h-20" />
       </main>
     </LayoutClient>
   );
